# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Vita API Documentation",
        "description": "petshop api",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "local server"
        },
        {
            "url": "https://8009.freelancedeveloper.site",
            "description": "remote server"
        }
    ],
    "paths": {
...
        "/api/v1/blocks": {
            "get": {
                "tags": [
                    "Blocks"
                ],
                "summary": "Get all blocks",
                "operationId": "fb220b334b9dd6b08f401c28caf3028e",
                "responses": {
                    "200": {
                        "description": "Blocks found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/BlockSchema"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Blocks"
                ],
                "summary": "Create a new block",
                "operationId": "576cceb20d0c7fc8e16fb2925b4cd38b",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title",
                                    "start_date"
                                ],
                                "properties": {
                                    "title": {
                                        "description": "The title of the block",
                                        "type": "string"
                                    },
                                    "details": {
                                        "description": "Additional details about the block",
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "repeat_every": {
                                        "description": "Repeat every specified interval",
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "repeat_type": {
                                        "description": "Type of repetition (e.g., day, week)",
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "repeat_on": {
                                        "description": "Days to repeat on (if applicable)",
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "nullable": true
                                    },
                                    "start_date": {
                                        "description": "Must be greater than or equal to the now",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-06-05"
                                    },
                                    "end_date": {
                                        "description": "The end date of the block",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-06-05",
                                        "nullable": true
                                    },
                                    "from_time": {
                                        "description": "Start time of the block",
                                        "type": "string",
                                        "format": "time",
                                        "example": "12:15"
                                    },
                                    "to_time": {
                                        "description": "End time of the block",
                                        "type": "string",
                                        "format": "time",
                                        "example": "15:00"
                                    },
                                    "end_on": {
                                        "description": "You can only provide one of the following: end_on, end_after",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-06-05",
                                        "nullable": true
                                    },
                                    "end_after": {
                                        "description": "You can only provide one of the following: end_on, end_after",
                                        "type": "integer",
                                        "example": 5,
                                        "nullable": true
                                    },
                                    "color": {
                                        "description": "Color associated with the block",
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Block created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/BlockSchema"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/blocks/filter/{date}": {
            "get": {
                "tags": [
                    "Blocks"
                ],
                "summary": "Get all blocks by date",
                "operationId": "af682c1b921c7a93bc78e38b9e459981",
                "parameters": [
                    {
                        "name": "date",
                        "in": "path",
                        "description": "The date of the resource",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Sort the blocks by date",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Blocks found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/BlockSchema"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "BlockSchema": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string"
                    },
                    "details": {
                        "type": "string"
                    },
                    "start_date": {
                        "type": "string",
                        "format": "date"
                    },
                    "from_time": {
                        "type": "string",
                        "example": "09:00:00"
                    },
                    "to_time": {
                        "type": "string",
                        "example": "12:00:00"
                    },
                    "end_date": {
                        "type": "string",
                        "format": "date"
                    },
                    "color": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "tasks": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/BlockTaskSchema"
                        }
                    }
                },
                "type": "object"
            },
            "BlockTaskSchema": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "block_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string"
                    },
                    "details": {
                        "type": "string"
                    },
                    "completed": {
                        "type": "boolean"
                    },
                    "urgent": {
                        "type": "boolean"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    }
}